+++ interface/src/avatar/MyCharacterController.cpp
@@ -108,7 +108,7 @@ void MyCharacterController::updateShapeIfNecessary() {
 
 std::vector<btTransform> MyCharacterController::getWorldCollisionTransforms() const {
     std::vector<btTransform> transforms;
-    for (int i = 0; i < _detailedCollisions._rigidBodies.size(); i++) {
+    for (int i = 0; i < (int)_detailedCollisions._rigidBodies.size(); i++) {
         if (_detailedCollisions._rigidBodies[i]._rigidBody) {
             transforms.push_back(_detailedCollisions._rigidBodies[i]._rigidBody->getWorldTransform());
         } else {
